Add toggle component and update application interface
Adding a new toggle component to the user interface, alongside updates to the main application interface. The changes include the addition of a @radix-ui/react-toggle dependency, as this library is necessary to make the toggle component work as intended. Additional UI components have also been added into the main page interface, along with some array objects representing filters to demonstrate their usage in the application. These additions and adjustments contribute to enhancing the interactivity and user experience of the application.