Add Importance Toggle to NewPostForm
Added a new "importance" field to the NewPostForm component. This allows users to set the urgency level of a post when creating a new one. A Toggle UI component was implemented for user to select one of the predefined time options. The selected option is stored in the "importance" state using the 'useEffect' hook, which monitors any changes in the activeToggle state and updates the 'importance' field value.