Modify NewPostForm for better data structure and validation
Altered 'importance' type to be an object in the NewPostForm schema to better manage and validate the data. Created ImportanceType interface and updated 'timeOptions' array to contain objects of ImportanceType. This allows that each timeOption carries both a 'name' and 'urgency' level. This aids in data manipulation and more precise validation of form values, improving the overall data integrity.