"Integrate Supabase Client in NewPostForm"
Supabase Client is incorporated in the NewPostForm component to provide DB functionality. Import statements for SupabaseClient and PostInsert/PostRow are added to support this. Supabase is passed as a prop in Page component. The onSubmit function in NewPostForm is updated to send data to the supabase instance. PostInsert type is used to ensure the correct data format. This allows for saving form data to the Database and further enhancing data management. The definition for the new type, PostInsert, is added in the supabase.ts file.