"Added UI switch component and integrated it in NewPostForm"
A new UI Switch component was created, and a corresponding import statement was added to the NewPostForm component. This improves the user experience by allowing more interactive and intuitive form controls. Additionally, the '@radix-ui/react-switch' package was added to the project dependencies to support the creation of the Switch component. For the same reason, package-lock.json was also updated. Further, NewPostForm was enhanced with a new military field using the Switch component, allowing users to toggle and submit this particular information.