Refactor post filtering logic and type definition
Improved structure and efficiency of post-filtering logic across NewPostForm, ComboBox and Page components, with a particular emphasis on Page. Changes include addition of eslint-disabled comments on "react-hooks/exhaustive-deps" rule to prevent unnecessary re-renders, introduction of 'FilterType' interface to strengthen typing, and extensive alteration of filtering functions and effects to streamline and render more flexible the setting of active filters and options, as well as the posts' backend retrieval process. These updates are meant to boost performance and enhance code readability and maintainability.