Update post loading feature and cleanup in Page component
The changes have been applied in order to improve the loading of content and minimize the redundancy in the code. The program fetches a limited number of posts initially and allows users to load more posts when required, hence updating the experience and loading speed. The repetitive code that used various calculations, effect hooks, and context state updates has been excluded ensuring a more readable and cleaner code. The superfluous Loader Icon was also introduced to indicate loading status.