Implement localization in Header and MainPage components
Introduced localization to enable multi-language support in the application. Language dictionaries were updated and utilized in the Header and MainPage components. This allows for dynamic translation of certain texts within these components based on the selected language by the user.  Adjusted the 'newAd' and 'amEhad' text to pull from the updated dictionaries in their respective component.