Implement support for multiple languages
Added support for multiple languages ('en', 'ru', 'he') to the application, and adapted the site directionality to change dynamically based on the selected language. 'lang' property was added to the ContextType interface and passed as a prop to components. The language-specific layout formatting which was formerly positioned in home function has been moved to MainPage component for more efficient code structure. This allows for better language handling and a smoother user experience for non-English speakers. Updated dropdown menu with correct links to switch between languages.