Optimize text display on MainPage for better accessibility
The positioning of the title text on MainPage is now dynamic, not hardcoded, improving overall user experience. The change shifts the text to the center, enhancing readability and visual aesthetics. Allowing the display of 'form.title' and 'form.toHelpTitle' from translations dictionary also adds to the application's accessibility by making translation to multiple languages easier.