Add ReCAPTCHA verification for improved security
ReCAPTCHA verification functionality has been added to the new post form. The RECAPTCHA site key has been added to the environmental variables and utilized in the NewPostForm component. Extra dependencies are added to package.json accordingly. The verification process happens in the backend in a new file, ServerActions.ts using axios to send a POST request to the siteverify API endpoint from google. This addition increases the overall security of the website by preventing bots and spam submissions.