Improve new post form security with ReCAPTCHA
Added ReCAPTCHA verification to the new post form to increase the security of the website. This change will prevent bots and spam submissions. The RECAPTCHA site key is included in the environmental variables. The verification process happens on the backend, utilizing axios for a POST request to Google's siteverify API.