The given commit message does not match the diff.
Based on the diff, a suitable commit message could be:

Refine the 'housingText' filter description in Hebrew Dictionary

Amended the 'housingText' filter description in the Hebrew dictionary to provide a more precise meaning. Enhanced the description to specify that housing solutions are for those in need. This change helps users with better understanding of filter usage.