Move postCard component out of ui directory
The postCard component was initially placed in the ui directory, which was not the most suitable place for it as it is not a general utility. It has been moved out of the ui directory and into the components directory where it belongs. Also, the import paths have been updated to reflect this change. The aim of this modification is to uphold proper file organization and enhance code readability.