Enhance languages menu in Header component
Made changes to the languages dropdown menu in the Header component.
The Links for 'עברית', 'English' and 'Русский' now occupy the full width of the dropdown item. This increases visibility and makes it easier to select a language.
Minor layout refactor to remove an unnecessary line and aligned the 'DropdownLanguage' menu properly.