Add ESLint exception, remove unused useState and useEffect in Filters
Added exception for 'react/no-unescaped-entities' rule in ESLint for Terms and Privacy pages in order to handle specific design intents. The previously utilized useState and useEffect hooks in the Filter component have been removed after the review highlighted them as redundant. The height calculation for the subfilterRef element was unnecessary and therefore commented out. The className set to manage transitions has been simplified and modified to depend on the selectedFilter value directly. This refactor enhances readability and maintainability.