"Add right to left layout support for Hebrew language
Modified the direction of the layout on terms of use and privacy policy pages based on the language selection. When a user selects Hebrew as their preferred language, the page layout will switch from the default left to right to right to left orientation. This change improves localization support and enhances user experience for readers of languages that are read from right to left, starting with Hebrew."