"Refactor layout file to support multilanguage
layout.tsx file was refactor to support multiple languages. The layout file was moved from /src/app to /src/app/[lang] and the path reference to the globals.css was also updated. generateStaticParams function was moved to just below the imports for better readability and the console.log on line 49 was removed to clean up the code."